+++ database/doc.go
@@ -1,0 +1,2 @@
+// Package database provides common functions for database images.
+package database
