+++ database/postgres/doc.go
@@ -1,0 +1,2 @@
+// Package postgres provides setup for spinning up a postgres container.
+package postgres
