+++ elasticsearch/doc.go
@@ -1,2 +1,2 @@
-// Package testcontainerselasticsearch provides setup for spinning up an elasticsearch container.
-package testcontainerselasticsearch
+// Package elasticsearch provides setup for spinning up an elasticsearch container.
+package elasticsearch
