--- sql/doc.go
@@ -1,2 +1,0 @@
-// Package testcontainersql provides common functions for sql images.
-package testcontainersql
