--- sql/postgres/doc.go
@@ -1,2 +1,0 @@
-// Package testcontainerspostgres provides setup for spinning up a postgres container.
-package testcontainerspostgres
